Changed "Advanced" to "Mail" + Added labels above inputs -  Removed Placeholders
