+++ js/admin/src/components/AdminNav.js
@@ -67,11 +67,11 @@ export default class AdminNav extends Component {
       description: app.translator.trans('core.admin.nav.extensions_text')
     }));
 
-    items.add('advanced', AdminLinkButton.component({
-      href: app.route('advanced'),
-      icon: 'cog',
-      children: app.translator.trans('core.admin.nav.advanced_button'),
-      description: app.translator.trans('core.admin.nav.advanced_text')
+    items.add('mail', AdminLinkButton.component({
+      href: app.route('mail'),
+      icon: 'envelope',
+      children: app.translator.trans('core.admin.nav.mail_button'),
+      description: app.translator.trans('core.admin.nav.mail_text')
     }))
 
     return items;
