+++ js/admin/src/initializers/routes.js
@@ -3,7 +3,7 @@ import BasicsPage from 'flarum/components/BasicsPage';
 import PermissionsPage from 'flarum/components/PermissionsPage';
 import AppearancePage from 'flarum/components/AppearancePage';
 import ExtensionsPage from 'flarum/components/ExtensionsPage';
-import AdvancedPage from 'flarum/components/AdvancedPage';
+import MailPage from 'flarum/components/MailPage';
 
 /**
  * The `routes` initializer defines the admin app's routes.
@@ -17,6 +17,6 @@ export default function(app) {
     'permissions': {path: '/permissions', component: PermissionsPage.component()},
     'appearance': {path: '/appearance', component: AppearancePage.component()},
     'extensions': {path: '/extensions', component: ExtensionsPage.component()},
-    'advanced': {path: '/advanced', component: AdvancedPage.component()}
+    'mail': {path: '/mail', component: MailPage.component()}
   };
 }
