+++ less/admin/MailPage.less
@@ -1,4 +1,4 @@
-.AdvancedPage {
+.MailPage {
   padding: 20px 0;
 
   @media @desktop-up {
@@ -19,10 +19,14 @@
   }
 }
 
-.AdvancedPage-mailSettings-input {
+.MailPage-MailSettings-input {
+
+  label {
+    margin-bottom: 5px;
+  }
 
   .FormControl {
-    margin-bottom: 1px;
+    margin-bottom: 7px;
   }
 
   :last-child {
