+++ less/admin/app.less
@@ -7,4 +7,4 @@
 @import "EditGroupModal.less";
 @import "ExtensionsPage.less";
 @import "AppearancePage.less";
-@import "AdvancedPage.less";
+@import "MailPage.less";
